requirements.txt
llama-cpp-python==0.2.56
langchain==0.1.13
langchain-community==0.0.29
openpyxl==3.1.2
pandas==2.2.1
numpy==1.26.4
python-dotenv==1.0.1
xlwings==0.30.13